+++ base_account_budget/models/account_analytic_account.py
@@ -3,7 +3,7 @@
 #
 #    Cybrosys Technologies Pvt. Ltd.
 #
-#    Copyright (C) 2019-TODAY Cybrosys Technologies(<https://www.cybrosys.com>)
+#    Copyright (C) 2023-TODAY Cybrosys Technologies(<https://www.cybrosys.com>)
 #    Author: Cybrosys Techno Solutions(<https://www.cybrosys.com>)
 #
 #    You can modify it under the terms of the GNU LESSER
@@ -23,6 +23,10 @@ from odoo import fields, models
 
 
 class AccountAnalyticAccount(models.Model):
+    """Inherits the AccountAnalytic model to add new budget line field that
+    connect with the budget line modules"""
     _inherit = "account.analytic.account"
 
-    budget_line = fields.One2many('budget.lines', 'analytic_account_id', 'Budget Lines')
+    budget_line = fields.One2many('budget.lines',
+                                  'analytic_account_id',
+                                  'Budget Lines')