+++ base_accounting_kit/models/account_account.py
@@ -3,7 +3,7 @@
 #
 #    Cybrosys Technologies Pvt. Ltd.
 #
-#    Copyright (C) 2022-TODAY Cybrosys Technologies(<https://www.cybrosys.com>)
+#    Copyright (C) 2023-TODAY Cybrosys Technologies(<https://www.cybrosys.com>)
 #    Author: Cybrosys Techno Solutions(<https://www.cybrosys.com>)
 #
 #    You can modify it under the terms of the GNU LESSER
@@ -19,17 +19,17 @@
 #    If not, see <http://www.gnu.org/licenses/>.
 #
 #############################################################################
-
-import time
-from odoo import api, models, fields, _
+from odoo import api, fields, models
 from odoo.tools.misc import get_lang
-from odoo.exceptions import UserError
 
 
 class CashFlow(models.Model):
+    """Inherits the account.account model to add additional functionality and
+     fields to the account"""
     _inherit = 'account.account'
 
     def get_cash_flow_ids(self):
+        """Returns a list of cashflows for the account"""
         cash_flow_id = self.env.ref('base_accounting_kit.account_financial_report_cash_flow0')
         if cash_flow_id:
             return [('parent_id.id', '=', cash_flow_id.id)]
@@ -38,12 +38,13 @@ class CashFlow(models.Model):
 
     @api.onchange('cash_flow_type')
     def onchange_cash_flow_type(self):
+        """Onchange the cash flow type of the account that will be updating
+        the account_ids values"""
         for rec in self.cash_flow_type:
             # update new record
             rec.write({
                 'account_ids': [(4, self._origin.id)]
             })
-
         if self._origin.cash_flow_type.ids:
             for rec in self._origin.cash_flow_type:
                 # remove old record
@@ -51,25 +52,29 @@ class CashFlow(models.Model):
 
 
 class AccountCommonReport(models.Model):
+    """Inherits the Account report model to add special fields and functions"""
     _inherit = "account.report"
     _description = "Account Common Report"
 
-    company_id = fields.Many2one('res.company', string='Company', required=True, readonly=True, default=lambda self: self.env.company)
+    company_id = fields.Many2one('res.company', string='Company',
+                                 required=True, readonly=True,
+                                 default=lambda self: self.env.company)
     journal_ids = fields.Many2many(
         comodel_name='account.journal',
         string='Journals',
         required=True,
         default=lambda self: self.env['account.journal'].search([('company_id', '=', self.company_id.id)]),
-        domain="[('company_id', '=', company_id)]",
-    )
+        domain="[('company_id', '=', company_id)]")
     date_from = fields.Date(string='Start Date')
     date_to = fields.Date(string='End Date')
     target_move = fields.Selection([('posted', 'All Posted Entries'),
                                     ('all', 'All Entries'),
-                                    ], string='Target Moves', required=True, default='posted')
+                                    ], string='Target Moves',
+                                   required=True, default='posted')
 
     @api.onchange('company_id')
     def _onchange_company_id(self):
+        """Onchange function based on the company and updated the journals"""
         if self.company_id:
             self.journal_ids = self.env['account.journal'].search(
                 [('company_id', '=', self.company_id.id)])
@@ -77,6 +82,7 @@ class AccountCommonReport(models.Model):
             self.journal_ids = self.env['account.journal'].search([])
 
     def _build_contexts(self, data):
+        """Builds the context information for the given data"""
         result = {}
         result['journal_ids'] = 'journal_ids' in data['form'] and data['form']['journal_ids'] or False
         result['state'] = 'target_move' in data['form'] and data['form']['target_move'] or ''
@@ -87,9 +93,12 @@ class AccountCommonReport(models.Model):
         return result
 
     def _print_report(self, data):
+        """Raise an error if the report comes checked """
         raise NotImplementedError()
 
     def check_report(self):
+        """Function to check if the report comes active models and related
+        values"""
         self.ensure_one()
         data = {}
         data['ids'] = self.env.context.get('active_ids', [])
@@ -101,24 +110,44 @@ class AccountCommonReport(models.Model):
 
 
 class AccountCommonJournalReport(models.TransientModel):
+    """Model used for creating the common journal report"""
     _name = 'account.common.journal.report'
     _description = 'Common Journal Report'
     _inherit = "account.report"
 
-    amount_currency = fields.Boolean('With Currency', help="Print Report with the currency column if the currency differs from the company currency.")
-
-    company_id = fields.Many2one('res.company', string='Company', required=True, readonly=True, default=lambda self: self.env.company)
+    section_main_report_ids = fields.Many2many(string="Section Of",
+                                               comodel_name='account.report',
+                                               relation="account_common_journal_report_section_rel",
+                                               column1="sub_report_id",
+                                               column2="main_report_id")
+    section_report_ids = fields.Many2many(string="Sections",
+                                          comodel_name='account.report',
+                                          relation="account_common_journal_report_section_rel",
+                                          column1="main_report_id",
+                                          column2="sub_report_id")
+    amount_currency = fields.Boolean(
+        'With Currency',
+        help="Print Report with the currency column if the currency differs "
+             "from the company currency.")
+    company_id = fields.Many2one('res.company', string='Company',
+                                 required=True, readonly=True,
+                                 default=lambda self: self.env.company)
     date_from = fields.Date(string='Start Date')
     date_to = fields.Date(string='End Date')
     target_move = fields.Selection([('posted', 'All Posted Entries'),
                                     ('all', 'All Entries'),
-                                    ], string='Target Moves', required=True, default='posted')
+                                    ], string='Target Moves',
+                                   required=True, default='posted')
 
     def pre_print_report(self, data):
+        """Pre-print the given data and that updates the amount
+        amount_currency value"""
         data['form'].update({'amount_currency': self.amount_currency})
         return data
 
     def check_report(self):
+        """Function to check if the report comes active models and related
+                values"""
         self.ensure_one()
         data = {}
         data['ids'] = self.env.context.get('active_ids', [])
@@ -129,6 +158,7 @@ class AccountCommonJournalReport(models.TransientModel):
         return self.with_context(discard_logo_check=True)._print_report(data)
 
     def _build_contexts(self, data):
+        """Builds the context information for the given data"""
         result = {}
         result['journal_ids'] = 'journal_ids' in data['form'] and data['form']['journal_ids'] or False
         result['state'] = 'target_move' in data['form'] and data['form']['target_move'] or ''
@@ -137,7 +167,3 @@ class AccountCommonJournalReport(models.TransientModel):
         result['strict_range'] = True if result['date_from'] else False
         result['company_id'] = data['form']['company_id'][0] or False
         return result
-
-
-
-