+++ base_accounting_kit/models/account_move.py
@@ -3,7 +3,7 @@
 #
 #    Cybrosys Technologies Pvt. Ltd.
 #
-#    Copyright (C) 2022-TODAY Cybrosys Technologies(<https://www.cybrosys.com>)
+#    Copyright (C) 2023-TODAY Cybrosys Technologies(<https://www.cybrosys.com>)
 #    Author: Cybrosys Techno Solutions(<https://www.cybrosys.com>)
 #
 #    You can modify it under the terms of the GNU LESSER
@@ -19,18 +19,17 @@
 #    If not, see <http://www.gnu.org/licenses/>.
 #
 #############################################################################
-
 from datetime import datetime
 import ast
 from dateutil.relativedelta import relativedelta
-
 from odoo import api, fields, models, _
-from odoo.addons.base.models import decimal_precision as dp
 from odoo.exceptions import UserError
 from odoo.tools import DEFAULT_SERVER_DATE_FORMAT as DF
 
 
 class AccountMove(models.Model):
+    """Inherits from the account.move model for adding the depreciation
+    field to the account"""
     _inherit = 'account.move'
 
     asset_depreciation_ids = fields.One2many('account.asset.depreciation.line',
@@ -38,18 +37,20 @@ class AccountMove(models.Model):
                                              string='Assets Depreciation Lines')
 
     def button_cancel(self):
+        """Button action to cancel the transfer"""
         for move in self:
             for line in move.asset_depreciation_ids:
                 line.move_posted_check = False
         return super(AccountMove, self).button_cancel()
 
     def post(self):
-
+        """Supering the post method to mapped the asset depreciation records"""
         self.mapped('asset_depreciation_ids').post_lines_and_close_asset()
         return super(AccountMove, self).action_post()
 
     @api.model
     def _refund_cleanup_lines(self, lines):
+        """Supering the refund cleanup lines to check the asset category """
         result = super(AccountMove, self)._refund_cleanup_lines(lines)
         for i, line in enumerate(lines):
             for name, field in line._fields.items():
@@ -59,20 +60,22 @@ class AccountMove(models.Model):
         return result
 
     def action_cancel(self):
+        """Action perform to cancel the asset record"""
         res = super(AccountMove, self).action_cancel()
         self.env['account.asset.asset'].sudo().search(
             [('invoice_id', 'in', self.ids)]).write({'active': False})
         return res
 
     def action_post(self):
+        """Action used to post invoice"""
         result = super(AccountMove, self).action_post()
-
         for inv in self:
             context = dict(self.env.context)
             # Within the context of an invoice,
-            # this default value is for the type of the invoice, not the type of the asset.
-            # This has to be cleaned from the context before creating the asset,
-            # otherwise it tries to create the asset with the type of the invoice.
+            # this default value is for the type of the invoice, not the type
+            # of the asset. This has to be cleaned from the context before
+            # creating the asset,otherwise it tries to create the asset with
+            # the type of the invoice.
             context.pop('default_type', None)
             inv.invoice_line_ids.with_context(context).asset_create()
         return result
@@ -96,6 +99,7 @@ class AccountInvoiceLine(models.Model):
 
     @api.depends('asset_category_id', 'move_id.invoice_date')
     def _get_asset_date(self):
+        """Returns the asset_start_date and the asset_end_date of the Asset"""
         for record in self:
             record.asset_mrr = 0
             record.asset_start_date = False
@@ -117,6 +121,7 @@ class AccountInvoiceLine(models.Model):
                     record.asset_end_date = end_date.strftime(DF)
 
     def asset_create(self):
+        """Create function for the asset and its associated properties"""
         for record in self:
             if record.asset_category_id:
                 vals = {
@@ -141,6 +146,8 @@ class AccountInvoiceLine(models.Model):
 
     @api.depends('asset_category_id')
     def onchange_asset_category_id(self):
+        """On change function based on the category and its updates the
+        account status"""
         if self.move_id.move_type == 'out_invoice' and self.asset_category_id:
             self.account_id = self.asset_category_id.account_asset_id.id
         elif self.move_id.move_type == 'in_invoice' and self.asset_category_id:
@@ -148,12 +155,15 @@ class AccountInvoiceLine(models.Model):
 
     @api.onchange('product_id')
     def _onchange_uom_id(self):
+        """Onchange function for product that's call the UOM compute function
+         and the asset category function"""
         result = super(AccountInvoiceLine, self)._compute_product_uom_id()
         self.onchange_asset_category_id()
         return result
 
     @api.depends('product_id')
     def _onchange_product_id(self):
+        """Onchange product values and it's associated with the move types"""
         vals = super(AccountInvoiceLine, self)._compute_price_unit()
         if self.product_id:
             if self.move_id.move_type == 'out_invoice':
@@ -163,6 +173,8 @@ class AccountInvoiceLine(models.Model):
         return vals
 
     def _set_additional_fields(self, invoice):
+        """The function adds additional fields that based on the invoice
+        move types"""
         if not self.asset_category_id:
             if invoice.type == 'out_invoice':
                 self.asset_category_id = self.product_id.product_tmpl_id.deferred_revenue_category_id.id
@@ -172,12 +184,14 @@ class AccountInvoiceLine(models.Model):
         super(AccountInvoiceLine, self)._set_additional_fields(invoice)
 
     def get_invoice_line_account(self, type, product, fpos, company):
+        """"It returns the invoice line and callback"""
         return product.asset_category_id.account_asset_id or super(
             AccountInvoiceLine, self).get_invoice_line_account(type, product,
                                                                fpos, company)
 
     @api.model
     def _query_get(self, domain=None):
+        """Used to add domain constraints to the query"""
         self.check_access_rights('read')
 
         context = dict(self._context or {})