+++ base_accounting_kit/wizard/trial_balance.py
@@ -3,7 +3,7 @@
 #
 #    Cybrosys Technologies Pvt. Ltd.
 #
-#    Copyright (C) 2022-TODAY Cybrosys Technologies(<https://www.cybrosys.com>)
+#    Copyright (C) 2023-TODAY Cybrosys Technologies(<https://www.cybrosys.com>)
 #    Author: Cybrosys Techno Solutions(<https://www.cybrosys.com>)
 #
 #    You can modify it under the terms of the GNU LESSER
@@ -28,6 +28,16 @@ class AccountBalanceReport(models.TransientModel):
     _name = 'account.balance.report'
     _description = 'Trial Balance Report'
 
+    section_report_ids = fields.Many2many(string="Sections",
+                                          comodel_name='account.report',
+                                          relation="account_balance_report_section_rel",
+                                          column1="main_report_id",
+                                          column2="sub_report_id")
+    section_main_report_ids = fields.Many2many(string="Section Of",
+                                               comodel_name='account.report',
+                                               relation="account_balance_report_section_rel",
+                                               column1="sub_report_id",
+                                               column2="main_report_id")
     name = fields.Char(string="Trial Balance", default="Trial Balance", required=True, translate=True)
     journal_ids = fields.Many2many('account.journal',
                                    'account_balance_report_journal_rel',